Refactor cockpit history helpers
